requirements.txt
coverage==4.4.2
Django==2.0.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
ldap3==2.4
pyasn1==0.4.2
python-decouple==3.1
pytz==2017.3
